pushed contents of Profile.jsx involving delete and datafetching to components folder
