+++ client/src/pages/Profile.jsx
@@ -1,78 +1,11 @@
 import { useState, useEffect } from 'react';
+import ProfileDataFetcher from '../components/ProfileDataFetcher';
+import ProfileDeleteHandler from '../components/ProfileDeleteHandler';
 
 const Profile = ({ token }) => {
     const [user, setUser] = useState({});
     const [posts, setPosts] = useState([]);
     const [replies, setReplies] = useState([]);
-
-    useEffect(() => {
-
-        const fetchData = async () => {
-            try {
-                // Fetch user data
-                const userResponse = await fetch("/api/user", {
-                    headers: {
-                        "Authorization" : `Bearer ${token}`
-                    }
-                });
-                const userData = await userResponse.json();
-                setUser(userData);
-
-                // Fetch posts
-                const postsResponse = await fetch("/api/posts", {
-                  headers: {
-                    "Authorization": `Bearer ${token}`
-                  }
-                });
-                const postsData = await postsResponse.json();
-                setPosts(postsData);
-
-                // Fetch replies
-                const repliesResponse = await fetch("/api/replies", {
-                  headers: {
-                    "Authorization": `Bearer ${token}`
-                  }
-                });
-                const repliesData = await repliesResponse.json();
-                setReplies(repliesData);
-
-            } catch (error) {
-                console.error('Error fetching user data:', error);
-            }
-        }
-
-        fetchData();
-    }, [token]);
-
-    const handleDeletePost = async (postId) => {
-      try {
-        await fetch(`/api/posts/${postId}`, {
-          method: 'DELETE',
-          headers: {
-            "Authorization": `Bearer ${token}`
-          }
-        });
-
-        setPosts(posts.filter((post) => post.id !== token.userId));
-      } catch (error) {
-        console.error('Error deleting post:', error);
-      }
-    };
-
-    // const handleDeleteReply = async (replyId) => {
-    //   try {
-    //     await fetch(`/api/replies/${replyId}`, {
-    //       method: 'DELETE',
-    //       headers: {
-    //         "Authorization": `Bearer ${token}`
-    //       }
-    //     });
-
-    //     setReplies(replies.filter((reply) => reply.id !== replyId));
-    //   } catch (error) {
-    //     console.error('Error deleting reply:', error);
-    //   }
-    // };
 
     const formatUser = (user) => {
         return (
@@ -90,7 +23,7 @@ const Profile = ({ token }) => {
               <h3>{post.title}</h3>
               <p>{post.content}</p>
               <img src={post.postImg} alt={`Post image ${post.id}`} />
-              <h5>Likes: {post.likes - post.dilikes}</h5>   
+              {/* <h5>Likes: {post.likes - post.dilikes}</h5>    */}
               {/* Need to change likes and dislikes to integer values to count em */}
             </ul>
           </div>
@@ -110,12 +43,18 @@ const Profile = ({ token }) => {
         );
     };  
 
+const handleDeletePost = (postId) => {
+  onDeletePost(postId);
+}
+    
   return (
     <div>
         {/* User Data */}
         {user && formatUser(user)}
         {/* Create Route for Post page off of button  */}
         <button>Create a Post!</button>
+
+        <ProfileDataFetcher token={token} setUser={setUser} setPosts={setPosts} setReplies={setReplies} />
         {/* Render Posts */}
         {posts.map((post) => {
             return formatPosts(post);
@@ -124,6 +63,8 @@ const Profile = ({ token }) => {
         {replies.map((reply) => {
             return formatReplies(reply);
         })}
+
+        <ProfileDeleteHandler token={token} posts={posts} setPosts={setPosts} onDeletePost={handleDeletePost} replies={replies} setReplies={setReplies} />
     </div>
   );
 }
